Update script for copying files to a GCP disk
 - check if VM exists before attempting to create it
 - update configure_ssh to just call `gcloud compute config-ssh`
 - extend retry time for SSH configuration (can sometimes take 30s+)
 - add a wrapper for `gcloud compute ssh` to auto-retry SSH commands,
   these can occasionally fail for no apparent reason but will succeed
   if retried
 - run `sync` after copying files and properly unmount the disk
 - make comment formatting more consistent

+++ gcp_deployment/copy_files_to_gcp_disk.sh
@@ -30,7 +30,8 @@ declare -r disk_name="${2}"
 declare -r zone="${3}"
 declare -r src="${4}"
 declare -r dest="${5}"
-declare -r new_disk="${6}"
+# not readonly as it might be updated below
+declare new_disk="${6}"
 
 check_source_exists() {
     # Check if the local source directory actually exists
@@ -49,7 +50,8 @@ check_source_exists() {
 is_response_not_empty() {
     # Check if a gcloud response is empty or not
     # 
-    # $1 = response text
+    # Arguments:
+        # $1 = response text
     #
     # Return value: 0 if argument is NOT empty, 1 if it IS empty
     [[ -n "${1}" ]]
@@ -58,12 +60,25 @@ is_response_not_empty() {
 does_disk_exist() {
     # Check if a disk with the given label already exists
     #
-    # $1 = label of the disk
-    # $2 = zone to filter on
+    # Arguments:
+    #   $1 = label of the disk
+    #   $2 = zone to filter on
     #
     # Return value: 0 if disk exists, 1 if not
 
     resp=$(gcloud compute disks list --zones="${2}" --filter=name="${1}" 2> /dev/null)
+    is_response_not_empty "${resp}"
+}
+
+does_vm_exist() {
+    # Check if a VM with the given name already exists
+    #
+    # Arguments:
+    #   $1 = name of the VM
+    #   $2 = zone to filter on
+    #
+    # Return value: 0 if VM exists, 1 if not
+    resp=$(gcloud compute instances list --zones="${2}" --filter=name="${1}" 2> /dev/null)
     is_response_not_empty "${resp}"
 }
 
@@ -76,14 +91,45 @@ configure_ssh() {
     #   $1 = VM name
     #
     # Return value: 0 if successful, 1 if not
-    declare num_retries=3
-    declare retry_delay=2
+    declare num_retries=4
+    declare retry_delay=10
     declare return_code=1
 
     for (( i=1; i<=num_retries; i++ )) 
     do
         echo "> Generating SSH keys and logging into VM..."
-        if ! gcloud compute ssh "${1}" --command "ls /" --quiet >/dev/null 2>&1
+        if ! gcloud compute config-ssh 2>&1
+        then
+            echo "    (retry #${i}/${num_retries})"
+            sleep "${retry_delay}"
+        else
+            return_code=0
+            break
+        fi
+    done
+
+    return ${return_code}
+}
+
+run_ssh_command() {
+    # runs a command in a VM over SSH, with auto-retry if it fails since this
+    # happens occasionally for no apparent reason
+    #
+    # Arguments:
+    #   $1 = VM name
+    #   $2 = SSH command string
+    #   $3 = message to display describing command
+    #
+    # Return value: 0 if successful, 1 if not
+
+    declare num_retries=4
+    declare retry_delay=5
+    declare return_code=1
+
+    for (( i=1; i<=num_retries; i++ )) 
+    do
+        echo "> ${3}..."
+        if ! gcloud compute ssh "${1}" --command "${2}"
         then
             echo "    (retry #${i}/${num_retries})"
             sleep "${retry_delay}"
@@ -105,9 +151,9 @@ format_disk() {
     #   
     #   Return value: ignored (should exit on error)
 
-
-    echo "> Formatting the disk..."
-    gcloud compute ssh "${1}" --command "sudo /sbin/mkfs.ext4 -q /dev/disk/by-id/google-${2}"
+    run_ssh_command "${1}" "sudo /sbin/mkfs.ext4 -q /dev/disk/by-id/google-${2}" "Formatting the disk"
+    # echo "> Formatting the disk..."
+    # gcloud compute ssh "${1}" --command "sudo /sbin/mkfs.ext4 -q /dev/disk/by-id/google-${2}"
 }
 
 scp_files_to_disk() {
@@ -123,12 +169,10 @@ scp_files_to_disk() {
 
     declare -r mountpoint="/mnt/temp"
     # Mount the drive at a fixed location
-    echo "> Mounting drive at ${mountpoint} in the VM"
-    gcloud compute ssh "${1}" --command "sudo mkdir -p ${mountpoint} && sudo mount /dev/disk/by-id/google-${2} ${mountpoint}"
+    run_ssh_command "${1}" "sudo mkdir -p ${mountpoint} && sudo mount /dev/disk/by-id/google-${2} ${mountpoint}" "Mounting drive at ${mountpoint} in the VM"
 
     # Create the path structure on the remote disk
-    echo "> Creating remote directory structure ${mountpoint}/${4}"
-    gcloud compute ssh "${1}" --command "sudo mkdir -p ${mountpoint}/${4} && sudo chmod -R a+rwx ${mountpoint}"
+    run_ssh_command "${1}" "sudo mkdir -p ${mountpoint}/${4} && sudo chmod -R a+rwx ${mountpoint}" "Creating remote dir structure ${mountpoint}/${4}"
 
     # Copy the files 
     declare local_sz
@@ -136,10 +180,17 @@ scp_files_to_disk() {
     echo "> Copying a total of ${local_sz}MB of files to remote..."
     gcloud compute scp --recurse "${3}"/* "${1}:${mountpoint}/${4}"
     echo "> Finished copying files!"
+    run_ssh_command "${1}" "sudo sync && sudo umount ${mountpoint}" "Unmounting disk"
+    echo "> Disk unmounted"
 }
 
-echo "> Creating a VM instance called ${vm_name}..."
-gcloud compute instances create "${vm_name}" --image-family=debian-11 --image-project=debian-cloud --machine-type=f1-micro --network-tier=STANDARD 2>/dev/null
+if ! does_vm_exist "${vm_name}" "${zone}"
+then
+    echo "> Creating a VM instance called ${vm_name}..."
+    gcloud compute instances create "${vm_name}" --image-family=debian-11 --image-project=debian-cloud --machine-type=f1-micro --network-tier=STANDARD
+else
+    echo "> Using existing VM ${vm_name}"
+fi
 
 if ! configure_ssh "${vm_name}"
 then
@@ -152,7 +203,16 @@ does_disk_exist "${disk_name}" "${zone}"
 
 # using device-name here allows the disk to be easily accessed under that for mounting/formatting
 echo "> Attaching disk ${disk_name}..."
-gcloud compute instances attach-disk "${vm_name}" --disk "${disk_name}" --device-name="${disk_name}" 2>/dev/null
+if ! gcloud compute instances attach-disk "${vm_name}" --disk "${disk_name}" --device-name="${disk_name}"
+then
+    # probably means already attached from a previous attempt"
+    echo "> Failed to attach disk, attempting to detach it..."
+    gcloud compute instances detach-disk "${vm_name}" --disk "${disk_name}"
+    # assume it wasn't formatted on the previous attempt
+    new_disk="true"
+    echo "> Re-attaching disk ${disk_name}..."
+    gcloud compute instances attach-disk "${vm_name}" --disk "${disk_name}" --device-name="${disk_name}"
+fi
 
 if [[ "${new_disk}" = "true" ]]
 then
@@ -166,10 +226,10 @@ scp_files_to_disk "${vm_name}" "${disk_name}" "${src}" "${dest}"
 
 # detach the disk from the VM
 echo "> Detaching disk..."
-gcloud compute instances detach-disk "${vm_name}" --disk "${disk_name}" 2>/dev/null
+gcloud compute instances detach-disk "${vm_name}" --disk "${disk_name}"
 
 # dispose of VM
 echo "> Deleting VM instance ${vm_name}..."
-gcloud compute instances delete "${vm_name}" --quiet 2>/dev/null
+gcloud compute instances delete "${vm_name}" --quiet 
 
 echo "> Completed successfully!"
